member router with error handling

+++ infrastructure/library_repository.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm import Session
 from infrastructure.models import Book, Member
 from uuid import UUID
 from datetime import datetime, timezone
+from shared.exceptions import NotFoundException, AlreadyBorrowedException
 
 class LibraryRepository:
     db: Session
@@ -17,8 +18,10 @@ class LibraryRepository:
         return new_book
     
     def get_book_by_id(self, book_id: int):
-        #TODO check existense
-        book = self.db.query(Book).filter(Book.id == book_id).first()
+        book = self.db.query(Book).filter(Book.book_id == book_id).first()
+        if not Book:
+            raise NotFoundException(f"Book with id: {book_id} is not found.")
+        
         return book
     
     def get_all_books(self):
@@ -26,9 +29,10 @@ class LibraryRepository:
     
     def update_book(self, book_id: int, updated_book: Book):
         #TODO updated_book validation logic
-        book_to_update = self.db.query(Book).filter(Book.id == book_id).first()
+        book_to_update = self.db.query(Book).filter(Book.book_id == book_id).first()
 
-        #TODO add existense check
+        if not book_to_update:
+            raise NotFoundException(f"Book with id: {book_id} is not found.")
 
         self.db.merge(updated_book)
         self.db.commit()
@@ -36,18 +40,24 @@ class LibraryRepository:
         return updated_book
 
     def delete_book_by_id(self, book_id: int):
-        book_to_delete = self.db.query(Book).filter(Book.id == book_id).first()
+        book_to_delete = self.db.query(Book).filter(Book.book_id == book_id).first()
 
-        #TODO add existense check
+        if not book_to_delete:
+            raise NotFoundException(f"Book with id: {book_id} is not found.")
 
         self.db.delete(book_to_delete)
         self.db.commit()
 
     def borrow_book(self, book_id: int, member_id: UUID):
-        book_to_borrow = self.db.query(Book).filter(Book.id == book_id).first()
-        member_to_assign = self.db.query(Member).filter(Member.id == member_id).first()
+        book_to_borrow = self.db.query(Book).filter(Book.book_id == book_id).first()
+        member_to_assign = self.db.query(Member).filter(Member.member_id == member_id).first()
 
-        #TODO validation logic for both
+        if not book_to_borrow:
+            raise NotFoundException(f"Book with id: {book_id} is not found.")
+        if not member_to_assign:
+            raise NotFoundException(f"Member with id: {member_id} doesn't exist.")
+        elif book_to_borrow.is_borrowed:
+            raise AlreadyBorrowedException(f"Book with id: {book_id} is already borrowed.")
 
         book_to_borrow.is_borrowed = True
         book_to_borrow.borrowed_by = member_id
@@ -57,7 +67,7 @@ class LibraryRepository:
         self.db.commit()
 
     def return_book(self, book_id: int):
-        book_to_return = self.db.query(Book).filter(Book.id == book_id).first()
+        book_to_return = self.db.query(Book).filter(Book.book_id == book_id).first()
 
         #TODO check if is_borrowed
 
@@ -76,28 +86,33 @@ class LibraryRepository:
         return new_member
     
     def get_member_by_id(self, member_id: int):
-        #TODO check existense
-        book = self.db.query(Member).filter(Member.id == member_id).first()
-        return book
+        member = self.db.query(Member).filter(Member.member_id == member_id).first()
+
+        if not member:
+            raise NotFoundException(f"Member with id: {member_id} is not found.")
+        
+        return member
     
     def get_all_members(self):
         return self.db.query(Member).all()
     
     def update_member(self, member_id: UUID, updated_member: Book):
         #TODO updated_member validation logic
-        member_to_update = self.db.query(Member).filter(Member.id == member_id).first()
+        member_to_update = self.db.query(Member).filter(Member.member_id == member_id).first()
 
-        #TODO add existense check
-
+        if not member_to_update:
+            raise NotFoundException(f"Member with id: {member_id} is not found.")
+        
         self.db.merge(updated_member)
         self.db.commit()
-        self.db.refresh(updated_member)
-        return updated_member
+        self.db.refresh(member_to_update)
+        return member_to_update
 
     def delete_member_by_id(self, member_id: UUID):
-        member_to_delete = self.db.query(Member).filter(Member.id == member_id).first()
+        member_to_delete = self.db.query(Member).filter(Member.member_id == member_id).first()
 
-        #TODO add existense check
+        if not member_to_delete:
+            raise NotFoundException(f"Member with id: {member_id} is not found.")
 
         self.db.delete(member_to_delete)
         self.db.commit()
